url_shortner/requirements.txt
django==4.1.2
django-rest-framework==0.1.0
django-cors-headers==3.13.0